+++ templates/repo/activity.tmpl
@@ -108,11 +108,8 @@
 						{{.i18n.Tr "repo.activity.git_stats_and_deletions" }}
 						<strong class="text red">{{.i18n.Tr (TrN .i18n.Lang .Activity.Code.Deletions "repo.activity.git_stats_deletion_1" "repo.activity.git_stats_deletion_n") .Activity.Code.Deletions }}</strong>.
 					</div>
-					<div class="ui attached segment" id="app">
-						<script type="text/javascript">
-						var ActivityTopAuthors = {{Json .ActivityTopAuthors | SafeJS}};
-						</script>
-						<activity-top-authors :data="activityTopAuthors" />
+					<div class="ui attached segment">
+						<div id="repo-activity-top-authors-chart"></div>
 					</div>
 				</div>
 			{{end}}
@@ -126,7 +123,7 @@
 			<div class="list">
 				{{range .Activity.PublishedReleases}}
 					<p class="desc">
-						<div class="ui green label">{{$.i18n.Tr "repo.activity.published_release_label"}}</div>
+						<span class="ui green label">{{$.i18n.Tr "repo.activity.published_release_label"}}</span>
 						{{.TagName}}
 						{{if not .IsTag}}
 							<a class="title" href="{{$.RepoLink}}/src/{{.TagName | EscapePound}}">{{.Title | RenderEmoji}}</a>
@@ -145,7 +142,7 @@
 			<div class="list">
 				{{range .Activity.MergedPRs}}
 					<p class="desc">
-						<div class="ui purple label">{{$.i18n.Tr "repo.activity.merged_prs_label"}}</div>
+						<span class="ui purple label">{{$.i18n.Tr "repo.activity.merged_prs_label"}}</span>
 						#{{.Index}} <a class="title" href="{{$.RepoLink}}/pulls/{{.Index}}">{{.Issue.Title | RenderEmoji}}</a>
 						{{TimeSinceUnix .MergedUnix $.Lang}}
 					</p>
@@ -161,7 +158,7 @@
 			<div class="list">
 				{{range .Activity.OpenedPRs}}
 					<p class="desc">
-						<div class="ui green label">{{$.i18n.Tr "repo.activity.opened_prs_label"}}</div>
+						<span class="ui green label">{{$.i18n.Tr "repo.activity.opened_prs_label"}}</span>
 						#{{.Index}} <a class="title" href="{{$.RepoLink}}/pulls/{{.Index}}">{{.Issue.Title | RenderEmoji}}</a>
 						{{TimeSinceUnix .Issue.CreatedUnix $.Lang}}
 					</p>
@@ -177,7 +174,7 @@
 			<div class="list">
 				{{range .Activity.ClosedIssues}}
 					<p class="desc">
-						<div class="ui red label">{{$.i18n.Tr "repo.activity.closed_issue_label"}}</div>
+						<span class="ui red label">{{$.i18n.Tr "repo.activity.closed_issue_label"}}</span>
 						#{{.Index}} <a class="title" href="{{$.RepoLink}}/issues/{{.Index}}">{{.Title | RenderEmoji}}</a>
 						{{TimeSinceUnix .ClosedUnix $.Lang}}
 					</p>
@@ -193,7 +190,7 @@
 			<div class="list">
 				{{range .Activity.OpenedIssues}}
 					<p class="desc">
-						<div class="ui green label">{{$.i18n.Tr "repo.activity.new_issue_label"}}</div>
+						<span class="ui green label">{{$.i18n.Tr "repo.activity.new_issue_label"}}</span>
 						#{{.Index}} <a class="title" href="{{$.RepoLink}}/issues/{{.Index}}">{{.Title | RenderEmoji}}</a>
 						{{TimeSinceUnix .CreatedUnix $.Lang}}
 					</p>
@@ -212,7 +209,7 @@
 			<div class="list">
 				{{range .Activity.UnresolvedIssues}}
 					<p class="desc">
-						<div class="ui green label">{{$.i18n.Tr "repo.activity.unresolved_conv_label"}}</div>
+						<span class="ui green label">{{$.i18n.Tr "repo.activity.unresolved_conv_label"}}</span>
 						#{{.Index}}
 						{{if .IsPull}}
 						<a class="title" href="{{$.RepoLink}}/pulls/{{.Index}}">{{.Title | RenderEmoji}}</a>
