+++ templates/repo/view_file.tmpl
@@ -131,13 +131,3 @@
 		</div>
 	</div>
 </div>
-
-<script>
-function submitDeleteForm() {
-	var message = prompt("{{.i18n.Tr "repo.delete_confirm_message"}}\n\n{{.i18n.Tr "repo.delete_commit_summary"}}", "Delete '{{.TreeName}}'");
-	if (message != null) {
-		$("#delete-message").val(message);
-		$("#delete-file-form").submit()
-	}
-}
-</script>
