+++ web_src/js/components/ActivityHeatmap.vue
@@ -70,4 +70,3 @@ export default {
   },
 };
 </script>
-<style scoped/>
