+++ web_src/less/_repository.less
@@ -415,10 +415,6 @@
             opacity: var(--opacity-disabled);
             cursor: default;
           }
-
-          #delete-file-form {
-            display: inline-block;
-          }
         }
       }
